refactor: move more methods
