+++ pkg/skaffold/deploy/helm/deploy30.go
@@ -17,22 +17,30 @@ limitations under the License.
 package helm
 
 import (
+	"bufio"
 	"bytes"
 	"context"
 	"fmt"
 	"io"
+	"io/ioutil"
+	"os"
 	"strings"
 
+	"github.com/blang/semver"
+	shell "github.com/kballard/go-shellquote"
+
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/constants"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/label"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/types"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/instrumentation"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/manifest"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output"
 	olog "github.com/GoogleContainerTools/skaffold/pkg/skaffold/output/log"
-	latestV1 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest/v1"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
-	"github.com/blang/semver"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/yaml"
 )
 
 // Deployer30 deploys workflows using the helm CLI less than 3.1
@@ -40,7 +48,7 @@ type Deployer30 struct {
 	*Deployer3
 }
 
-func NewDeployer30(ctx context.Context, cfg Config, labeller *label.DefaultLabeller, h *latestV1.HelmDeploy, hv semver.Version) (*Deployer30, error) {
+func NewDeployer30(ctx context.Context, cfg Config, labeller *label.DefaultLabeller, h *latest.HelmDeploy, hv semver.Version) (*Deployer30, error) {
 	d3, err := NewBase(ctx, cfg, labeller, h, hv)
 	if err != nil {
 		return nil, err
@@ -124,6 +132,114 @@ func (h *Deployer30) Deploy(ctx context.Context, out io.Writer, builds []graph.A
 	return nil
 }
 
+// deployRelease deploys a single release
+func (h *Deployer30) deployRelease(ctx context.Context, out io.Writer, releaseName string, r latest.HelmRelease, builds []graph.Artifact, valuesSet map[string]bool, helmVersion semver.Version, chartVersion string) ([]types.Artifact, error) {
+	var err error
+	opts := h.getInstallOpts(releaseName, r, helmVersion, chartVersion)
+
+	var installEnv []string
+	if h.enableDebug {
+		if h.bV.LT(helm31Version) {
+			return nil, fmt.Errorf("debug requires at least Helm 3.1 (current: %v)", h.bV)
+		}
+		var binary string
+		if binary, err = osExecutable(); err != nil {
+			return nil, fmt.Errorf("cannot locate this Skaffold binary: %w", err)
+		}
+		opts.postRenderer = binary
+
+		var buildsFile string
+		if len(builds) > 0 {
+			var cleanup func()
+			buildsFile, cleanup, err = writeBuildArtifacts(builds)
+			if err != nil {
+				return nil, fmt.Errorf("could not write build-artifacts: %w", err)
+			}
+			defer cleanup()
+		}
+
+		cmdLine := h.generateSkaffoldDebugFilter(buildsFile)
+
+		// need to include current environment, specifically for HOME to lookup ~/.kube/config
+		env := util.EnvSliceToMap(util.OSEnviron(), "=")
+		env["SKAFFOLD_CMDLINE"] = shell.Join(cmdLine...)
+		env["SKAFFOLD_FILENAME"] = h.configFile
+		installEnv = util.EnvMapToSlice(env, "=")
+	}
+
+	opts.namespace, err = h.releaseNamespace(r)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := h.exec(ctx, ioutil.Discard, false, nil, getArgs(releaseName, opts.namespace)...); err != nil {
+		output.Yellow.Fprintf(out, "Helm release %s not installed. Installing...\n", releaseName)
+
+		opts.upgrade = false
+		opts.flags = h.Flags.Install
+	} else {
+		if r.UpgradeOnChange != nil && !*r.UpgradeOnChange {
+			olog.Entry(ctx).Infof("Release %s already installed...", releaseName)
+			return []types.Artifact{}, nil
+		} else if r.UpgradeOnChange == nil && r.RemoteChart != "" {
+			olog.Entry(ctx).Infof("Release %s not upgraded as it is remote...", releaseName)
+			return []types.Artifact{}, nil
+		}
+	}
+
+	// Only build local dependencies, but allow a user to skip them.
+	if !r.SkipBuildDependencies && r.ChartPath != "" {
+		olog.Entry(ctx).Info("Building helm dependencies...")
+
+		if err := h.exec(ctx, out, false, nil, "dep", "build", r.ChartPath); err != nil {
+			return nil, userErr("building helm dependencies", err)
+		}
+	}
+
+	// Dump overrides to a YAML file to pass into helm
+	if len(r.Overrides.Values) != 0 {
+		overrides, err := yaml.Marshal(r.Overrides)
+		if err != nil {
+			return nil, userErr("cannot marshal overrides to create overrides values.yaml", err)
+		}
+
+		if err := ioutil.WriteFile(constants.HelmOverridesFilename, overrides, 0666); err != nil {
+			return nil, userErr(fmt.Sprintf("cannot create file %q", constants.HelmOverridesFilename), err)
+		}
+
+		defer func() {
+			os.Remove(constants.HelmOverridesFilename)
+		}()
+	}
+
+	if r.Packaged != nil {
+		chartPath, err := h.packageChart(ctx, r)
+		if err != nil {
+			return nil, userErr("cannot package chart", err)
+		}
+
+		opts.chartPath = chartPath
+	}
+
+	args, err := h.installArgs(r, builds, valuesSet, opts)
+	if err != nil {
+		return nil, userErr("release args", err)
+	}
+
+	err = h.exec(ctx, out, r.UseHelmSecrets, installEnv, args...)
+	if err != nil {
+		return nil, userErr("install", err)
+	}
+
+	b, err := h.getReleaseManifest(ctx, releaseName, opts.namespace)
+	if err != nil {
+		return nil, userErr("get release", err)
+	}
+
+	artifacts := parseReleaseManifests(opts.namespace, bufio.NewReader(&b))
+	return artifacts, nil
+}
+
 // Render generates the Kubernetes manifests and writes them out
 func (h *Deployer30) Render(ctx context.Context, out io.Writer, builds []graph.Artifact, offline bool, filepath string) error {
 	instrumentation.AddAttributesToCurrentSpanFromContext(ctx, map[string]string{
@@ -186,4 +302,17 @@ func (h *Deployer30) Render(ctx context.Context, out io.Writer, builds []graph.A
 	}
 
 	return manifest.Write(renderedManifests.String(), filepath, out)
+}
+
+func (h *Deployer30) generateSkaffoldDebugFilter(buildsFile string) []string {
+	args := []string{"filter", "--debugging", "--kube-context", h.kubeContext}
+	if len(buildsFile) > 0 {
+		args = append(args, "--build-artifacts", buildsFile)
+	}
+	args = append(args, h.Flags.Global...)
+
+	if h.kubeConfig != "" {
+		args = append(args, "--kubeconfig", h.kubeConfig)
+	}
+	return args
 }
